// apps/client/src/data/wearables.ts
export type WearableSlot =
  | 'head'
  | 'body'
  | 'face'
  | 'eyes'
  | 'handLeft'
  | 'handRight'
  | 'pet'
  | 'background'
  | 'none';